🚚 Move settings back
