+++ lamindb/core/__init__.py
@@ -5,7 +5,16 @@ Settings & context:
 .. autosummary::
    :toctree: .
 
+   Settings
+   subsettings
    Context
+
+Artifact loaders:
+
+.. autosummary::
+   :toctree: .
+
+   loaders
 
 Data loaders:
 
@@ -29,9 +38,7 @@ from lamin_utils import logger
 from lamin_utils._inspect import InspectResult
 
 from .. import errors as exceptions
-from . import datasets, types
+from . import datasets, loaders, subsettings, types
 from ._context import Context
 from ._mapped_collection import MappedCollection
-
-# backward compat
 from ._settings import Settings
