+++ lamindb/core/_sync_git.py
@@ -7,7 +7,7 @@ from lamin_utils import logger
 from lamindb_setup import settings as setup_settings
 from lamindb_setup.core.hashing import hash_code
 
-from ..models._settings import sanitize_git_repo_url, settings
+from ..core._settings import sanitize_git_repo_url, settings
 
 
 class BlobHashNotFound(SystemExit):
