+++ tests/core/test_artifact.py
@@ -17,6 +17,7 @@ import pandas as pd
 import pytest
 import spatialdata as sd
 import yaml  # type: ignore
+from lamindb.core._settings import settings
 from lamindb.core.loaders import load_fcs, load_to_memory, load_tsv
 from lamindb.core.storage._zarr import identify_zarr_type, write_adata_zarr
 from lamindb.core.storage.paths import (
@@ -29,7 +30,6 @@ from lamindb.errors import (
     IntegrityError,
     InvalidArgument,
 )
-from lamindb.models._settings import settings
 from lamindb.models.artifact import (
     check_path_is_child_of_root,
     data_is_anndata,
